% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_description}
\alias{extract_description}
\title{Get query description.}
\usage{
extract_description(lines)
}
\arguments{
\item{lines}{Character vector giving the query docstring.}
}
\value{
description Character vector.
}
\description{
Extract the query description from the docstring.
}

