% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_name}
\alias{extract_name}
\title{Get query name.}
\usage{
extract_name(string)
}
\arguments{
\item{string}{Character vector giving the query.}
}
\value{
query_name Character string.
}
\description{
Extract the name of the query from the docstring.
}

