% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sql}
\alias{extract_sql}
\title{Get SQL query.}
\usage{
extract_sql(lines)
}
\arguments{
\item{lines}{Character vector giving the full query.}
}
\value{
sql Character string.
}
\description{
Extract the SQL query.
}

