% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_query}
\alias{get_query}
\title{Get query.}
\usage{
get_query(filename)
}
\arguments{
\item{filename}{Character string.}
}
\value{
Named list.
}
\description{
Create an anonymous query from a sql file. The
query is returned as a named list with a description
and SQL string suitable for passing to a DBIConnection
method.
}
\details{
The file should contain only a single query to be
loaded.
}
\examples{
cat("SELECT * FROM fruit",
    "WHERE color = 'green';",
    file = "query.sql",
    sep = "\\n")
query <- get_query("query.sql")
unlink("query.sql")

query$sql
# [1] "SELECT * FROM fruit\\nWHERE color = 'green'"

}

