% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_sql}
\alias{get_sql}
\title{Return the raw SQL.}
\usage{
get_sql(query)
}
\arguments{
\item{query}{Function.}
}
\value{
Character string.
}
\description{
Get the SQL stored in the query closure.
}
\examples{
\dontrun{
require(RSQLite)

writeLines("SELECT DISTINCT color FROM fruits;", con = "query.sql")
load_query("get_colors", "query.sql")
get_sql(get_colors)
# [1] "SELECT DISTINCT color FROM fruits"
}

}

