% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{is_name}
\alias{is_name}
\title{Check if the line contains a query name.}
\usage{
is_name(string)
}
\arguments{
\item{string}{Character string giving the query name.}
}
\value{
bool
}
\description{
Check if the line contains a query name.
}

