% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{is_prepared}
\alias{is_prepared}
\title{Check if the query contains parameters.}
\usage{
is_prepared(sql)
}
\arguments{
\item{sql}{Character string.}
}
\value{
bool
}
\description{
Check if the query contains parameters.
}

