% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{load_query}
\alias{load_query}
\title{Load query into an environment.}
\usage{
load_query(name, filename, env = parent.frame())
}
\arguments{
\item{name}{Character string. Name to be given to the query.}

\item{filename}{Character string.}

\item{env}{Environment to load the queries into. Default
is the calling environment.}
}
\description{
Create a query function inside an environment. The query
is added to the specified environment as the given name.
}
\examples{
# A package implementing DBI methods must be available. 
require(RSQLite)
cat("SELECT * FROM fruit",
    "WHERE name LIKE '\%berry'",
    file = "query.sql",
    sep = "\\n")

# Load into global environment.
load_query("get_berries", "query.sql")
unlink("query.sql")

conn <- dbConnect(SQLite(), ":memory:")
tbl <- data.frame(name = c("apple", "strawberry", "blackberry", "mango"),
                  color = c("green", "red", "black", "orange"))
success <- dbWriteTable(conn, "fruit", tbl)

get_berries(conn)
#         name color
# 1 strawberry   red
# 2 blackberry black

dbDisconnect(conn)

}
\seealso{
get_query
}

