% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{mk_callable}
\alias{mk_callable}
\title{Create a callable query object.}
\usage{
mk_callable(query)
}
\arguments{
\item{query}{List.}
}
\value{
Function.
}
\description{
Convert a named list, as returned by `parse_named_query`,
into a function with attached metadata.
}
\details{
A package defining DBI methods must be loaded prior to
calling this function.
}

