% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{parse_anon_query}
\alias{parse_anon_query}
\title{Parse a docstring and sql block.}
\usage{
parse_anon_query(block)
}
\arguments{
\item{block}{Character vector.}
}
\value{
List.
}
\description{
Parse a docstring and sql block.
}

