% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{parse_named_query}
\alias{parse_named_query}
\title{Parse a docstring with name and sql block.}
\usage{
parse_named_query(block)
}
\arguments{
\item{block}{Character vector.}
}
\value{
Named query.
}
\description{
Parse a docstring with name and sql block.
}

