% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{query_type}
\alias{query_type}
\title{Determine the type of query.}
\usage{
query_type(query_name)
}
\arguments{
\item{query_name}{Character string. The parsed query name.}
}
\value{
type Character string.
}
\description{
Check the query name for identifying information.
}

