% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{split_queries}
\alias{split_queries}
\title{Split slurped SQL into named-queries.}
\usage{
split_queries(lines)
}
\arguments{
\item{lines}{Character vector of SQL statements to split.}
}
\value{
queries List of queries.
}
\description{
Split slurped SQL into named-queries.
}

