% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_queries}
\alias{get_queries}
\title{Get queries.}
\usage{
get_queries(filename)
}
\arguments{
\item{filename}{Character string.}
}
\value{
List of named-queries.
}
\description{
Create named-queries from a sql file. Each query is
returned as a named list with a name, description and
SQL string suitable for passing to a DBIConnection
method.
}
\examples{
cat("-- name: get-fruits",
    "-- Get all fruits.",
    "SELECT * FROM fruit",
    "",
    "-- name: green-fruits",
    "-- Get green fruits.",
    "SELECT * FROM fruit",
    "WHERE color = 'green'",
    file = "queries.sql",
    sep = "\\n")
queries <- get_queries("queries.sql")
unlink("queries.sql")

names(queries)
# [1] "get-fruits"   "green-fruits"
queries$`green-fruits`$sql
# [1] "SELECT * FROM fruit\\nWHERE color = 'green'"

}

