% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{load_queries}
\alias{load_queries}
\title{Load queries into an environment.}
\usage{
load_queries(filename, env = parent.frame())
}
\arguments{
\item{filename}{Character string.}

\item{env}{Environment to load the queries into. Default
is the calling environment.}
}
\description{
Create query functions inside an environment. Multiple
queries are added to the specified environment, identified
by their names.
}
\seealso{
get_queries
}

