% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{load_query}
\alias{load_query}
\title{Load query into an environment.}
\usage{
load_query(name, filename, env = parent.frame())
}
\arguments{
\item{name}{Character string. Name to be given to the query.}

\item{filename}{Character string.}

\item{env}{Environment to load the queries into. Default
is the calling environment.}
}
\description{
Create a query function inside an environment. The query
is added to the specified environment as the given name.
}
\seealso{
get_query
}

